+++ tests/conftest.py
@@ -11,6 +11,7 @@ import betamax
 import pytest
 from betamax_serializers import pretty_json
 
+
 # pylint: disable=import-error,no-name-in-module
 if sys.version_info.major == 2:
     from urllib import quote_plus  # NOQA